requirements.txt
discord.py==2.6.4
fastapi==0.128.0
uvicorn[standard]==0.40.0
PyYAML==6.0.3
aiosqlite==0.22.1
python-dotenv==1.2.1
httpx==0.27.2
audioop-lts; python_version>='3.13'
